requirements.txt
gitdb2==2.0.6
GitPython==2.1.15
requests==2.22.0